requirements.txt
requests==2.21.0
urllib3==1.24.2
websockets==7.0
python-dotenv
